requirements.txt
# Основные зависимости
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
statsmodels>=0.13.0
prophet>=1.1.0
pmdarima>=1.8.5
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
tensorflow>=2.8.0
torch>=1.10.0
transformers>=4.15.0
optuna>=2.10.0
mlflow>=1.23.0
joblib>=1.1.0
tqdm>=4.62.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0
click>=8.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
python-dotenv>=0.19.0
requests>=2.26.0
aiohttp>=3.8.0
asyncio>=3.4.3

# Зависимости для тестирования
pytest>=6.2.5
pytest-cov>=2.12.0
pytest-benchmark>=3.4.0
hypothesis>=6.36.0
coverage>=6.2.0
pytest-asyncio>=0.15.0

# Зависимости для документации
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
nbsphinx>=0.8.0
myst-parser>=0.15.0

# Зависимости для линтинга и форматирования
black>=21.7b0
isort>=5.9.0
flake8>=3.9.0
mypy>=0.910
bandit>=1.7.0
safety>=1.10.0
pre-commit>=2.15.0

# Зависимости для безопасности
cryptography>=3.4.0
pyjwt>=2.3.0
itsdangerous>=2.0.1
bleach>=4.1.0
markupsafe>=2.0.1

# Зависимости для мониторинга и визуализации
prometheus-client>=0.16.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
kaleido>=0.2.1
psutil>=5.9.0
python-dateutil>=2.8.2

# Зависимости для распределенных вычислений
dask>=2022.1.0
distributed>=2022.1.0
ray>=2.0.0
ray[tune]>=2.0.0
ray[default]>=2.0.0

# Зависимости для баз данных
sqlalchemy>=1.4.0
pymongo>=4.0.0
redis>=4.0.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0

# Зависимости для форматов данных
pyarrow>=7.0.0
h5py>=3.6.0
pyyaml>=6.0.0
lxml>=4.9.0
openpyxl>=3.0.0
xlrd>=2.0.0 